Add checks in nearly-valid function pointers.
Morello has silly exception priorities (known bug) so will report a bus
error instead of a CHERI fault in some places.  If a jump is to
something that is both and not a valid capability for a jump and not a
valid MMU mapping, the MMU fault is reported as a higher priority.

This is not a security problem and should not block compliance, but to
permit this we also add a handler for SIGBUS in the CheriBSD
integration.  This potentially allows tests to pass if the MMU is the
*only* thing causing the trap and that *should* prevent certification.

To work around this, the new tests create function pointers that point
to things that are definitely valid executable MMU mappings but which
are not valid callable capabilities.